New check to test for a file's existence
The file.py check allows the agent to test for whether files exist in
the local file system, and if they do, emit a metric for how long they
have existed. This is particularly useful for files like Debian's
"package upgrades require a reboot" marker in /var/run/, or local puppet
facts meant to be kept around only for a short time.