Adding a check for filebeat
Filebeat (https://github.com/elastic/beats) is a deamon tailing and processing
log files, for example to send them to Logstash.

It maintains a _registry file_, in which it keeps track of which log files it's
currently tailing, and up to what point it has processed them.

This patch adds an agent check that leverages that registry file to gauge
filebat's progress; more specifically, it pushes how many bytes are currently
left to process for filebeat.

This metric would be useful for example to create a monitor alerting when
filebeat falls too far behind.

Added unit tests on the new check.