Making the filebeat check compatible with versions > 5.x, and making it able to use the HTTP profiler
This patch makes the filebeat check compatible with the new syntax Filebeat
started using in versions 5.x and later for its registry file.
It still maintains backward compatibility with the syntax from previous
versions.

This patch also makes it possible to push to Datadog the metrics Filebeat
exposes via its HTTP profiler, when started with the `--httpprof [HOST]:PORT`
option (see
https://www.elastic.co/guide/en/beats/filebeat/current/command-line-options.html)

It's worth noting since Filebeat reports a bunch of variables that are
counters from the last time it started, which by itself is not really interesting;
which is why we take care for these metrics to only push increments from
one check run to the next.

Added extensive tests on all of the above.