Fix connected but no active fullsync replication
The agent was throwing an exception when there is a connected cluster but no active fullsync replication. 

This fix will ignore fullsync replication metrics when there is no active  fullsync coordinator, which means a fullsync is not happening for this cluster.

In our use case riak is connected to multiple clusters but the fullsync replication is not active to all of those clusters. 

We tested and applied the proposed change and were able to get the metrics for the current active fullsync replication.