Add instructions to avoid confusing error
While the setup guide for the Developer Toolkit does talk about the extras repository,
it is formulated as "if you are working with". When a user is just following the
instructions from the datadog official documentation homepage she/he has no
idea about the extras repository. Make it clear this is needed.