helm: add helm check
Helm is a package manager for kubernetes.  More information
about helm is at https://helm.sh.  The command `helm list` shows
package status, such as revision number and if the package has failed
to install.

This adds a helm check that reports the revision number and if the
revision installed correctly.  It mirrors the numbering scheme
of https://github.com/sstarcher/helm-exporter to make migration
easier.

The only catch is it requires that the agent running the check have
permission to list kubernetes secrets, since that is how helm stores
metadata.