Update metrics.yaml
- https://a.cl.ly/yAubYQqA
- https://github.com/DataDog/integrations-extras/commit/edb8a1ceb0274e7bc371f8b6d6322f995456897d#diff-76f86e4b4fa718c1b24efa8d1b9ac080
- https://github.com/DataDog/integrations-extras/tree/e0720317b06a293fc61b61e474942e67573e2bb7/hbase_master
In above, we can see a _gen.rb that reads the _metrics.yml to generate the conf.yaml.example with the proper jmx metrics inside. This was all before we ever used a metrics.yaml for JMX. Not sure who owns _gen.rb though; never seen it before and now gone from master.