apollo: Update metric names and tags
Earlier this year as part of a rebranding we upgraded this integration to be
called "Apollo" rather than "Apollo Engine". We are now changing the metrics to
start with `apollo.operations` rather than `apollo.engine.operations` and
changing the tag name from `service` to `graph`.

Note that our infrastructure will continue using the old metric names and tags
for existing integrations until users choose to migrate over to the new
one, so `metric_to_check` uses a list instead of naming just one metric.