apollo: revert display name in CSV; update transition date (#729)
* apollo: Update transition date

PR #719 took longer than expected to make it to Datadog's infrastructure. On the
bright side, the message is more accurate when the release happens on the first
of the month :)

* Change metadata file display name back to apollo_engine

Apparently the attempt we've been making since March to remove our old product
name from these integrations isn't possible and we need to keep display_name
stuck on "Apollo Engine".