Update README after switching to Python binding
Primarily, the `base_command` goes away. There is still a mention of
`fdbcli`, but it is indicating no more than that our option names are
aligned with it (which is also a gentle hint that one can use `fdbcli`
to test that the configuration works). No change to instructions about
installing the FoundationDB client, as the CLI and libraries come in the
same package.