Add Fluent Bit internal metrics integration
Fluent Bit is a lightweight open-source logging and metrics processor and forwarder written in C. It is a CNCF sub-project under the umbrella of Fluentd.

Fluent Bit can be configured to expose its internal metrics in JSON and Prometheus formats, to allow pipeline monitoring: input, filter, and output plugins.

This integration scraps the Prometheus metrics endpoint.

* Removed unused import.