(Docs) Add github action to check PRs for invalid partials
This adds a check to find invalid/incomplete partials in our markdown.

We use `region` and `tabs` partials in our markdown to add functionality
to the docs site. However, our integrations docs are shared into both
the docs site and the Datadog app, and we need to add special HTML
comments around these partials so the app knows what to do with them.
Without these comments, tabs and region code breaks in the app.

This is a docs hackathon project.