go_pprof_scraper: make cumulative true by default
The Go runtime's implementation of "delta" profiles takes the difference
of every value type in a profile. This means in-use values in the heap
profile can be negative. This is different than what our client library
does, and it might be confusing for users. Make "cumulative" true by
default, and add a warning about negative values if it's set to false.