Add integration for collecting Go profiles (#1541)
* go_pprof_scraper: add new Go /debug/pprof/ profile scraper integration

Many services written in Go expose an HTTP endpoint (usually the default
/debug/pprof/) for collecting runtime profiles. These profiles can be
useful for debugging performance problems. Normally we encourage
customers to instrument their Go applications with the dd-trace-go
profiler client library. However, for external applications, maintaining
this instrumentation may be difficult. If this /debug/pprof/ endpoint is
available, we can still collect the profiles.

Add an integration to collect these profiles and upload them to Datadog.

* go_pprof_scraper: add env, runtime-id tags, fix timestamps

* go_pprof_scraper: make time format work for py 2 and 3

Also make the mock backend check the format to prevent regression

* go_pprof_scraper: post to UDS

* go_pprof_scraper: make cumulative true by default

The Go runtime's implementation of "delta" profiles takes the difference
of every value type in a profile. This means in-use values in the heap
profile can be negative. This is different than what our client library
does, and it might be confusing for users. Make "cumulative" true by
default, and add a warning about negative values if it's set to false.

* Apply suggestions from code review

Co-authored-by: Florent Clarret <florent.clarret@gmail.com>

* go_pprof_scraper: add installation to README.md, update metrics section

* go_pprof_scraper: migrate to manifest version 2.0.0

* go_pprof_scraper: rename "service_name" to "service" in config

* go_pprof_scraper: minor fixups in mock_backend.go

* go_pprof_scraper: skip generic tags check

The generic tags check fails when a "service" tag is present. But for
profiling, we require that profiles are tagged with a "service".

* go_pprof_scraper: run "ddev validate ci --fix"

* go_pprof_scraper: add code owner

* go_pprof_scraper: sync config example

* go_pprof_scraper: fix deprecated import

* go_pprof_scraper: add missing author info to manifest

* go_pprof_scraper: fix check field in manifest

* Apply suggestions from code review

Co-authored-by: Heston Hoffman <hestonhoffman@gmail.com>

* go_pprof_scraper: add documentation code owner info

* go_pprof_scraper: apply suggestions from code review

* Fix go_pprof_scraper display name in Azure pipline file

* Rename go_pprof_scraper in codecov.yml

* go_pprof_scraper: update version to 1.0.0

* go_pprof_scraper: fix name in spec

* go_pprof_scraper: make description shorter

* go_pprof_scraper: remove metrics section from manifest

* go_pprof_scraper: update agent version in service_checks.json

* go_pprof_scraper: fix app_id in manifest.json

* go_pprof_scraper: remove metrics checks from tests

* go_pprof_scraper: fix integratoin name in service_checks.json

* go_pprof_scraper: read actual agent config

Doing this doesn't work the way I expect when I set an agent config
value through an environment variable:

	apm_config = datadog_agent.get_config("apm_config")
	port = apm_config.get("receiver_port")

But this does:

	port = apm_config.get("apm_config.receiver_port")

So do it that way.

* go_pprof_scraper: remove unused import

* Update go_pprof_scraper/README.md

Co-authored-by: Florent Clarret <florent.clarret@gmail.com>

* go_pprof_scraper: update default duration, use hostname if scraping localhost

* go_pprof_scraper: udpate CHANGELOG.md

* go_pprof_scraper: update models

* go_pprof_scraper: update README.md

Co-authored-by: Florent Clarret <florent.clarret@gmail.com>
Co-authored-by: Heston Hoffman <hestonhoffman@gmail.com>