go_pprof_scraper: add profiler_version tag, fix min_collection_interval default (#1699)
* go_pprof_scraper: add profiler_version tag

The version in the tag is prefixed with "agent-integration" to distinguish Go
profiles uploaded by this integration from profiles uploaded by the dd-trace-go
client library.

* go_pprof_scraper: reset counts between tests

* go_pprof_scraper: default min_collection_interval to 0

I misunderstood min_collection_interval. From the docs[1]:

	Note: If the min_collection_interval is set to 30, it does not mean that
	the metric is collected every 30 seconds, but rather that it could be
	collected as often as every 30 seconds. The collector tries to run the
	check every 30 seconds but the check might need to wait in line,
	depending on how many integrations are enabled on the same Agent. Also
	if the check method takes more than 30 seconds to finish, the Agent
	skips execution until the next interval.

This check is intentionally long-running, and should start again as soon
as the previous iteration finishes. If the default value is 0, however,
the check will be run again as soon as possible.

[1] https://docs.datadoghq.com/developers/write_agent_check/?tab=agentv6v7#collection-interval

* go_pprof_scraper: update models

* go_pprof_scraper: set min_profiler_duration default to 1

Setting it to 0, or omitting the value, seems to pick up 15 seconds,
which is not what I want. So I set it to 1, meaning the check should run
again ASAP, and addded a note about setting the value for this
integration.

* go_pprof_scraper: update CHANGELOG.md

* Apply suggestions from code review

Co-authored-by: Andrew Zhang <andrew.zhang@datadoghq.com>

* go_pprof_scraper: add UDS fix to changelog

Co-authored-by: Andrew Zhang <andrew.zhang@datadoghq.com>