go_pprof_scraper: correctly handle URLs without trailing slashes (#1727)
* go_pprof_scraper: correctly handle URLs without trailing slashes

If the pprof_url parameter doesn't have a trailing slash, profiles won't
be collected from the expected URL. For example, given

	pprof_url: "http://1.2.3.4:6060/debug/pprof"

the CPU profile will be requested from

	"http://1.2.3.4:6060/debug/profile"

rather than

	"http://1.2.3.4:6060/debug/pprof/profile"

which is what it _should_ do. This is due to the way urljoin works. To
fix this, add a trailing slash if one isn't already there. This commit
also updates the pprof_url documentation to explain how profiles will be
collected.

* go_pprof_scraper: remove TODO

* Update go_pprof_scraper/CHANGELOG.md

Co-authored-by: Florent Clarret <florent.clarret@gmail.com>