Fix link (#2867)
This is currently rendering as the exact text that you see, and not
slotting in any value.  The intent of this looks like it was to
allow for any site to work, but it looks like we can do relative
paths now.

Brought up in the serverless support channel - https://dd.slack.com/archives/CKJAXN89L/p1766144821340449